+++ modules/serialize/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.1.0](https://github.com/aws/aws-encryption-sdk-javascript/compare/@aws-crypto/serialize@2.0.0...@aws-crypto/serialize@2.1.0) (2021-02-04)
+
+**Note:** Version bump only for package @aws-crypto/serialize
+
+
+
+
+
 # [2.0.0](https://github.com/aws/private-aws-encryption-sdk-javascript-staging/compare/@aws-crypto/serialize@1.7.0...@aws-crypto/serialize@2.0.0) (2020-09-25)